self implemented spherical k-means++ function
This program is a self implemented spherical k-means++ function in the C++ programming language. This algorithm is used in unsupervised learning to cluster similar data and the implementation assumes dense data & has been used on reuters-21578. It shows an excellent silhouette value of 0.82 subject to error in feature vector construction and takes in data in the form of normalized feature vectors, then randomly selects one of the feature vectors as an initial centre. The code then performs Farthest First Point (FFP) initialization to get a set of centres that are farthest from each other. Subsequently, it assigns clusters to each feature vector and computes normalized median of each cluster to find new centres. The process is repeated until there is no change in cluster assignment. This is a very concise implementation done at an atomic level to allow flexibility in various aspects such as computing similarity, distance etc. Here similarity is computed using cosine similarity which is essentially dot product since the feature vectors are normalized.